+++ commands/say.js
@@ -1,4 +1,5 @@
 exports.run = (client, message, args) => {
+  if (!message.guild) return;
   if (message.deletable) message.delete();
   message.channel.send(args.join(" "));
 };
